requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==8.10.0
openai==1.3.7
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
pydantic==2.5.0
python-dateutil==2.8.2
httpx==0.25.2
dateparser==1.1.8
pytz==2024.1
